Changed MovingBot.StartTime from float -> double

+++ Ported/DOTSBucketBrigade/Assets/Scripts/Components/MovingBot.cs
@@ -6,6 +6,6 @@ public struct MovingBot : IComponentData
 {
     public float2 StartPosition;
     public float2 TargetPosition;
-    public float StartTime;
+    public double StartTime;
     public int NewStateOnArrival;
 }
